+++ header/motor.h
@@ -6,15 +6,15 @@
 
 class Motor {
     private:
-        unsigned int _motorNumber;
-        bool         _polarity;
-        unsigned int _command;
-        unsigned int _dutyCycle;
+        byte                _motorNumber;
+        bool                _polarity;
+        unsigned int        _command;
+        unsigned int        _dutyCycle;
         
-        Adafruit_DCMotor *_AFMotor;
+        Adafruit_DCMotor    *_AFMotor;
         
     public: 
-        Motor(Adafruit_MotorShield *AFMS, unsigned int motorNumber, bool polarity);
+        Motor(Adafruit_MotorShield *AFMS, byte motorNumber, bool polarity);
         void driveCmd(unsigned int driveNum, unsigned int cmd);
         unsigned int getSpeed();
         void stop();
